# openapi.yml
openapi: "3.0.3"
info:
  title: Wishlist API
  version: "1.0"
  description: |
    API for managing wishlists and gift items.
    JSON request bodies must be one object, at most 1 MiB including trailing whitespace.
    Unknown properties, null values and NUL characters are rejected with 422.
    Titles must contain a non-whitespace character. Dates use YYYY-MM-DD, years 0001–9999.
    PUT replaces editable fields; PATCH changes only supplied fields; an empty PATCH is a no-op.
    Server failures return 500 with an Error object. Unknown routes and unsupported methods
    return JSON Error objects with 404 and 405 respectively.

servers:
  - url: http://localhost:8080/api/v1

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "409":
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: email already taken
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "password must contain 8 to 72 bytes"

  /auth/login:
    post:
      summary: Log in
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid email or password

        "422":
          $ref: "#/components/responses/ValidationError"

  /wishlists:
    get:
      summary: List my wishlists
      tags: [wishlists]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wishlist"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a wishlist
      tags: [wishlists]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wishlist"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: title is required

  /wishlists/{id}:
    get:
      summary: Get wishlist by id
      tags: [wishlists]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistWithItems"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wishlist not found

    put:
      summary: Replace wishlist
      tags: [wishlists]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wishlist"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wishlist not found

    patch:
      summary: Partially update wishlist
      tags: [wishlists]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wishlist"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wishlist not found

    delete:
      summary: Delete wishlist
      tags: [wishlists]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wishlist not found

  /wishlists/{id}/items:
    post:
      summary: Add item to wishlist
      tags: [items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: title is required

  /wishlists/{id}/items/{itemId}:
    put:
      summary: Replace item
      tags: [items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item not found

    patch:
      summary: Partially update item
      tags: [items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item not found

    delete:
      summary: Delete item
      tags: [items]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WishlistId"
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item not found

  /shared/{token}:
    get:
      summary: View wishlist by share token (no auth)
      tags: [public]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWishlistWithItems"
        "404":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wishlist not found

  /shared/{token}/items/{itemId}/reserve:
    post:
      summary: Reserve an item (no auth)
      tags: [public]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item not found
        "409":
          description: Already reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item already reserved

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    ValidationError:
      description: Invalid request JSON or field value
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized

  parameters:
    WishlistId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    AuthRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          description: "8 to 72 UTF-8 bytes (not character count)."
          example: "secret123"

    TokenResponse:
      type: object
      properties:
        token:
          type: string

    WishlistInput:
      type: object
      additionalProperties: false
      required: [title, event_date]
      properties:
        title:
          type: string
          example: Birthday 2026
        description:
          type: string
          example: Things I'd like to get
        event_date:
          type: string
          format: date
          example: "2026-06-15"

    WishlistPatch:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date

    Wishlist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date
        share_token:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    WishlistWithItems:
      allOf:
        - $ref: "#/components/schemas/Wishlist"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Item"

    PublicWishlist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time

    PublicWishlistWithItems:
      allOf:
        - $ref: "#/components/schemas/PublicWishlist"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Item"

    ItemInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          example: Mechanical keyboard
        description:
          type: string
          example: Cherry MX Brown switches
        url:
          type: string
          description: "Optional absolute URI, or an empty string. HTTP(S) URIs must have a hostname."
          example: https://example.com/keyboard
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
          description: "1 = nice to have, 5 = must have. Omitted on POST/PUT defaults to 3; explicit 0 is invalid."

    ItemPatch:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        url:
          type: string
          description: "Optional absolute URI, or an empty string. HTTP(S) URIs must have a hostname."
        priority:
          type: integer
          minimum: 1
          maximum: 5

    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wishlist_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        url:
          type: string
          description: "Optional absolute URI, or an empty string. HTTP(S) URIs must have a hostname."
        priority:
          type: integer
        is_reserved:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string